+++ packages/components/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orion-js/components
 
+## 4.0.32
+
+### Patch Changes
+
+- Updated dependencies [00f3a6f]
+  - @orion-js/mongodb@4.0.23
+  - @orion-js/dogs@4.0.25
+
+## 4.0.31
+
+### Patch Changes
+
+- Updated dependencies [e69c35d]
+  - @orion-js/mongodb@4.0.22
+  - @orion-js/dogs@4.0.24
+
+## 4.0.30
+
+### Patch Changes
+
+- Updated dependencies [27a0e4d]
+  - @orion-js/mongodb@4.0.21
+  - @orion-js/dogs@4.0.23
+
+## 4.0.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @orion-js/mongodb@4.0.20
+  - @orion-js/dogs@4.0.22
+
 ## 4.0.28
 
 ### Patch Changes