+++ packages/dogs/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @orion-js/dogs
 
+## 4.0.25
+
+### Patch Changes
+
+- Updated dependencies [00f3a6f]
+  - @orion-js/mongodb@4.0.23
+
+## 4.0.24
+
+### Patch Changes
+
+- Updated dependencies [e69c35d]
+  - @orion-js/mongodb@4.0.22
+
+## 4.0.23
+
+### Patch Changes
+
+- Updated dependencies [27a0e4d]
+  - @orion-js/mongodb@4.0.21
+
+## 4.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @orion-js/mongodb@4.0.20
+
 ## 4.0.21
 
 ### Patch Changes