+++ packages/file-manager/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @orion-js/file-manager
 
+## 4.0.32
+
+### Patch Changes
+
+- Updated dependencies [00f3a6f]
+  - @orion-js/mongodb@4.0.23
+  - @orion-js/paginated-mongodb@4.0.31
+
+## 4.0.31
+
+### Patch Changes
+
+- Updated dependencies [e69c35d]
+  - @orion-js/mongodb@4.0.22
+  - @orion-js/paginated-mongodb@4.0.30
+
+## 4.0.30
+
+### Patch Changes
+
+- Updated dependencies [27a0e4d]
+  - @orion-js/mongodb@4.0.21
+  - @orion-js/paginated-mongodb@4.0.29
+
+## 4.0.29
+
+### Patch Changes
+
+- Update mongodb use of rawCollection
+- Updated dependencies
+  - @orion-js/mongodb@4.0.20
+  - @orion-js/paginated-mongodb@4.0.28
+
 ## 4.0.28
 
 ### Patch Changes